Rename CryptoMarket rate provider to Notbank
